@@ -17,15 +17,15 @@
         "commands",
         "arguments"
     ],
-    "dependencies": {},
     "devDependencies": {
-        "@types/node": "^10.14.4",
+        "@types/node": "^18.0.0",
+        "discord.js": "^13.8.1",
         "discord.js-docgen": "github:discordjs/docgen",
-        "eslint": "^5.16.0",
-        "jsdoc": "^3.6.4",
-        "tslint": "^5.15.0",
+        "eslint": "^8.18.0",
+        "jsdoc": "^3.6.10",
+        "tslint": "^6.1.3",
         "tslint-config-typings": "^0.3.1",
-        "typescript": "^3.4.2"
+        "typescript": "^4.7.4"
     },
     "scripts": {
         "test": "npm run lint",
@@ -38,5 +38,11 @@
     "bugs": {
         "url": "https://github.com/discord-akairo/discord-akairo/issues"
     },
-    "homepage": "https://github.com/discord-akairo/discord-akairo"
+    "homepage": "https://github.com/discord-akairo/discord-akairo",
+    "packageManager": "yarn@3.2.1",
+    "optionalDependencies": {
+        "mongoose": "^6.4.1",
+        "sequelize": "^6.21.1",
+        "sqlite": "^4.1.1"
+    }
 }
