@@ -1,19 +1,24 @@
-const { Collection } = require('discord.js');
+import { Collection } from 'discord.js';
+import AkairoModule from '../struct/AkairoModule';
 
 /**
  * A group of modules.
- * @param {string} id - ID of the category.
- * @param {Iterable} [iterable] - Entries to set.
  * @extends {Collection}
  */
-class Category extends Collection {
-    constructor(id, iterable) {
+export default class Category<K extends string, V extends AkairoModule> extends Collection<K, V> {
+    /**
+     * ID of the category.
+     * @type {string}
+     */
+    public id: string;
+
+    /**
+     * @param {string} id - ID of the category.
+     * @param {Iterable} [iterable] - Entries to set.
+     */
+    public constructor(id: string, iterable?: Iterable<[K, V][]>) {
         super(iterable);
 
-        /**
-         * ID of the category.
-         * @type {string}
-         */
         this.id = id;
     }
 
@@ -21,7 +26,7 @@ class Category extends Collection {
      * Calls `reload()` on all items in this category.
      * @returns {Category}
      */
-    reloadAll() {
+    public reloadAll(): this {
         for (const m of Array.from(this.values())) {
             if (m.filepath) m.reload();
         }
@@ -33,7 +38,7 @@ class Category extends Collection {
      * Calls `remove()` on all items in this category.
      * @returns {Category}
      */
-    removeAll() {
+    public removeAll(): this {
         for (const m of Array.from(this.values())) {
             if (m.filepath) m.remove();
         }
@@ -45,9 +50,7 @@ class Category extends Collection {
      * Returns the ID.
      * @returns {string}
      */
-    toString() {
+    public toString(): string {
         return this.id;
     }
 }
-
-module.exports = Category;
