@@ -1,6 +1,6 @@
-const { Command } = require('../../src');
+import { Command } from '../../src';
 
-class ConditionalCommand extends Command {
+export default class ConditionalCommand extends Command {
     constructor() {
         super('condition');
     }
@@ -13,5 +13,3 @@ class ConditionalCommand extends Command {
         return message.util.reply('made you condition');
     }
 }
-
-module.exports = ConditionalCommand;
