@@ -1,8 +1,8 @@
 /* eslint-disable no-console */
 
-const { Command } = require('../../src');
+import { Command } from '../../src';
 
-class QCommand extends Command {
+export default class QCommand extends Command {
     constructor() {
         super('q', {
             aliases: ['q']
@@ -14,5 +14,3 @@ class QCommand extends Command {
         return this.handler.handleDirectCommand(message, '', command);
     }
 }
-
-module.exports = QCommand;
