@@ -1,38 +1,56 @@
-module.exports = {
-    // Core
-    AkairoClient: require('./struct/AkairoClient'),
-    AkairoHandler: require('./struct/AkairoHandler'),
-    AkairoModule: require('./struct/AkairoModule'),
-    ClientUtil: require('./struct/ClientUtil'),
+// Core
+export { default as AkairoClient } from './struct/AkairoClient';
+export * from './struct/AkairoClient';
+export { default as AkairoHandler } from './struct/AkairoHandler';
+export * from './struct/AkairoHandler';
+export { default as AkairoModule } from './struct/AkairoModule';
+export * from './struct/AkairoModule';
+export { default as ClientUtil } from './struct/ClientUtil';
+export * from './struct/ClientUtil';
 
-    // Commands
-    Command: require('./struct/commands/Command'),
-    CommandHandler: require('./struct/commands/CommandHandler'),
-    CommandUtil: require('./struct/commands/CommandUtil'),
-    Flag: require('./struct/commands/Flag'),
+// Commands
+export { default as Command } from './struct/commands/Command';
+export * from './struct/commands/Command';
+export { default as CommandHandler } from './struct/commands/CommandHandler';
+export * from './struct/commands/CommandHandler';
+export { default as CommandUtil } from './struct/commands/CommandUtil';
+export * from './struct/commands/CommandUtil';
+export { default as Flag } from './struct/commands/Flag';
+export * from './struct/commands/Flag';
 
-    // Arguments
-    Argument: require('./struct/commands/arguments/Argument'),
-    TypeResolver: require('./struct/commands/arguments/TypeResolver'),
+// Arguments
+export { default as Argument } from './struct/commands/arguments/Argument';
+export * from './struct/commands/arguments/Argument';
+export { default as TypeResolver } from './struct/commands/arguments/TypeResolver';
+export * from './struct/commands/arguments/TypeResolver';
 
-    // Inhibitors
-    Inhibitor: require('./struct/inhibitors/Inhibitor'),
-    InhibitorHandler: require('./struct/inhibitors/InhibitorHandler'),
+// Inhibitors
+export { default as Inhibitor } from './struct/inhibitors/Inhibitor';
+export * from './struct/inhibitors/Inhibitor';
+export { default as InhibitorHandler } from './struct/inhibitors/InhibitorHandler';
+export * from './struct/inhibitors/InhibitorHandler';
 
-    // Listeners
-    Listener: require('./struct/listeners/Listener'),
-    ListenerHandler: require('./struct/listeners/ListenerHandler'),
+// Listeners
+export { default as Listener } from './struct/listeners/Listener';
+export * from './struct/listeners/Listener';
+export { default as ListenerHandler } from './struct/listeners/ListenerHandler';
+export * from './struct/listeners/ListenerHandler';
 
-    // Providers
-    Provider: require('./providers/Provider'),
-    SequelizeProvider: require('./providers/SequelizeProvider'),
-    SQLiteProvider: require('./providers/SQLiteProvider'),
-    MongooseProvider: require('./providers/MongooseProvider'),
+// Providers
+export { default as Provider } from './providers/Provider';
+export * from './providers/Provider';
+export { default as SequelizeProvider } from './providers/SequelizeProvider';
+export * from './providers/SequelizeProvider';
+export { default as SQLiteProvider } from './providers/SQLiteProvider';
+export * from './providers/SQLiteProvider';
+export { default as MongooseProvider } from './providers/MongooseProvider';
+export * from './providers/MongooseProvider';
 
-    // Utilities
-    AkairoError: require('./util/AkairoError'),
-    Category: require('./util/Category'),
-    Constants: require('./util/Constants'),
-    Util: require('./util/Util'),
-    version: require('../package.json').version
-};
+// Utilities
+export { default as AkairoError } from './util/AkairoError';
+export * from './util/AkairoError';
+export { default as Category } from './util/Category';
+export * from './util/Category';
+export * as Constants from './util/Constants';
+export * as Util from './util/Util';
+export const version = require('../package.json').version;
