@@ -3,7 +3,7 @@
  * @param {string} type - Type of flag.
  * @param {any} [data={}] - Extra data.
  */
-class Flag {
+export default class Flag {
     constructor(type, data = {}) {
         this.type = type;
         Object.assign(this, data);
@@ -57,5 +57,3 @@ class Flag {
         return value instanceof Flag && value.type === type;
     }
 }
-
-module.exports = Flag;
